rename: Split mixin and styles
