+++ src/pages/ProjectText/ProjectTextarea/index.tsx
@@ -1,6 +1,6 @@
 import styled, { css } from 'styled-components';
 import { fontFamily } from '~/styles/mixins';
-import { mixin } from '../styles';
+import { mixin } from '../mixin';
 
 const ProjectTextarea = styled.textarea.attrs(() => ({
   placeholder: '텍스트를 입력하세요.',
