+++ src/pages/ProjectText/styles.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const Container = styled.div``;
+export const HeaderContainer = styled.div``;
+export const ContentsContainer = styled.div`
+  display: grid;
+  gap: 32px;
+`;
